Agregado swagger, usuarios y servicios de autentication con tests

+++ src/main/java/org/develop/web/funkos/FunkosWebController.java
@@ -1,14 +1,15 @@
 package org.develop.web.funkos;
 
+import jakarta.servlet.http.HttpSession;
 import jakarta.validation.Valid;
 import lombok.extern.slf4j.Slf4j;
-import org.develop.categorias.models.Categoria;
-import org.develop.categorias.services.CategoriasService;
-import org.develop.funkos.dto.FunkoCreateDto;
-import org.develop.funkos.dto.FunkoUpdateDto;
-import org.develop.funkos.models.Funko;
-import org.develop.funkos.repositories.FunkosRepository;
-import org.develop.funkos.services.FunkosService;
+import org.develop.rest.categorias.models.Categoria;
+import org.develop.rest.categorias.services.CategoriasService;
+import org.develop.rest.funkos.dto.FunkoCreateDto;
+import org.develop.rest.funkos.dto.FunkoUpdateDto;
+import org.develop.rest.funkos.models.Funko;
+import org.develop.rest.funkos.services.FunkosService;
+import org.develop.web.store.UserStore;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
@@ -19,7 +20,7 @@ import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
-import java.util.List;
+import java.util.Date;
 import java.util.Optional;
 
 
@@ -29,22 +30,61 @@ import java.util.Optional;
 public class FunkosWebController {
     private final FunkosService funkoService;
     private final CategoriasService categoriaService;
+    private final UserStore userSession;
 
     @Autowired
-    public FunkosWebController(FunkosService funkoService, CategoriasService categoriaService) {
+    public FunkosWebController(FunkosService funkoService, CategoriasService categoriaService, UserStore userSession) {
         this.funkoService = funkoService;
         this.categoriaService = categoriaService;
+        this.userSession = userSession;
+    }
+
+    @GetMapping("/login")
+    public String login(HttpSession session) {
+        log.info("Login GET");
+        if (isLoggedAndSessionIsActive(session)) {
+            log.info("Si está logueado volvemos al index");
+            return "redirect:/funkos";
+        }
+        return "funkos/login";
+    }
+
+    @PostMapping
+    public String login(@RequestParam("password") String password, HttpSession session, Model model) {
+        log.info("Login POST");
+        if ("pass".equals(password)) {
+            userSession.setLastLogin(new Date());
+            userSession.setLogged(true);
+            session.setAttribute("userSession", userSession);
+            session.setMaxInactiveInterval(1800);
+            return "redirect:/funkos";
+        } else {
+            return "funkos/login";
+        }
+    }
+
+    @GetMapping("/logout")
+    public String logout(HttpSession session) {
+        log.info("Logout GET");
+        session.invalidate();
+        return "redirect:/funkos";
     }
 
     @GetMapping(path = {"", "/", "/index", "/list"})
     public String index(
-                        Model model,
+                        HttpSession session, Model model,
                         @RequestParam(value = "search", required = false) Optional<String> search,
                         @RequestParam(defaultValue = "0") int page,
                         @RequestParam(defaultValue = "3") int size,
                         @RequestParam(defaultValue = "id") String sortBy,
                         @RequestParam(defaultValue = "asc") String direction
     ) {
+        // Comprobamos si está logueado
+        if (!isLoggedAndSessionIsActive(session)) {
+            log.info("No hay sesión o no está logueado volvemos al login");
+            return "redirect:/funkos/login";
+        }
+
         log.info("Index GET con parámetros search: " + search + ", page: " + page + ", size: " + size);
         Sort sort = direction.equalsIgnoreCase(Sort.Direction.ASC.name()) ? Sort.by(sortBy).ascending() : Sort.by(sortBy).descending();
 
@@ -86,7 +126,12 @@ public class FunkosWebController {
     }
 
     @GetMapping("/update/{id}")
-    public String updateForm(@PathVariable("id") Long id, Model model) {
+    public String updateForm(@PathVariable("id") Long id, Model model, HttpSession session) {
+        if (!isLoggedAndSessionIsActive(session)) {
+            log.info("No hay sesión o no está logueado volvemos al login");
+            return "redirect:/funkos/login";
+        }
+
         var categorias = categoriaService.findAll(null).stream().map(Categoria::getNombre).toList();
         Funko funko = funkoService.findById(id);
         FunkoUpdateDto funkoUpdateDto = FunkoUpdateDto.builder()
@@ -103,7 +148,7 @@ public class FunkosWebController {
     }
 
     @PostMapping("/update/{id}")
-    public String updateFunk(@PathVariable("id") Long id, @ModelAttribute FunkoUpdateDto funkoUpdateDto, BindingResult result, Model model) {
+    public String updateFunk(@PathVariable("id") Long id,@Valid @ModelAttribute("funko") FunkoUpdateDto funkoUpdateDto, BindingResult result, Model model) {
         if (result.hasErrors()) {
             var categorias = categoriaService.findAll(null).stream().map(Categoria::getNombre).toList();
             model.addAttribute("categorias", categorias);
@@ -117,30 +162,53 @@ public class FunkosWebController {
     }
 
     @GetMapping("/delete/{id}")
-    public String delete(@PathVariable("id") Long id) {
+    public String delete(@PathVariable("id") Long id, HttpSession session) {
+        if (!isLoggedAndSessionIsActive(session)) {
+            log.info("No hay sesión o no está logueado volvemos al login");
+            return "redirect:/funkos/login";
+        }
         funkoService.deleteById(id);
         return "redirect:/funkos";
     }
 
     @GetMapping("/update-image/{id}")
-    public String showUpdateImageForm(@PathVariable("id") Long funkoId, Model model) {
+    public String showUpdateImageForm(@PathVariable("id") Long funkoId, Model model, HttpSession session) {
+
+        if (!isLoggedAndSessionIsActive(session)) {
+            log.info("No hay sesión o no está logueado volvemos al login");
+            return "redirect:/funkos/login";
+        }
+
         Funko funko = funkoService.findById(funkoId);
         model.addAttribute("funko", funko);
         return "funkos/update-image";
     }
 
     @PostMapping("/update-image/{id}")
-    public String updateProductImage(@PathVariable("id") Long funkoId, @RequestParam("imagen") MultipartFile imagen) {
+    public String updateFunkImage(@PathVariable("id") Long funkoId, @RequestParam("imagen") MultipartFile imagen) {
         log.info("Update POST con imagen");
         funkoService.updateImage(funkoId, imagen);
         return "redirect:/funkos";
     }
 
     @GetMapping("/details/{id}")
-    public String details(@PathVariable("id") Long id, Model model) {
+    public String details(@PathVariable("id") Long id, Model model, HttpSession session) {
         log.info("Details GET");
+
+        if (!isLoggedAndSessionIsActive(session)) {
+            log.info("No hay sesión o no está logueado volvemos al login");
+            return "redirect:/funkos/login";
+        }
+
         Funko funko = funkoService.findById(id);
         model.addAttribute("funko", funko);
         return "funkos/details";
+    }
+
+
+    private boolean isLoggedAndSessionIsActive(HttpSession session) {
+        log.info("Comprobando si está logueado");
+        UserStore sessionData = (UserStore) session.getAttribute("userSession");
+        return sessionData != null && sessionData.isLogged();
     }
 }
